chore: convert SpinButton to index based stories approach
