+++ packages/react-components/react-spinbutton/src/stories/SpinButton/SpinButtonBounds.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { SpinButton } from '../index';
+import { SpinButton } from '@fluentui/react-spinbutton';
 import { Label } from '@fluentui/react-label';
 import { useId } from '@fluentui/react-utilities';
 import { tokens } from '@fluentui/react-theme';
@@ -17,23 +17,26 @@ const useLayoutStyles = makeStyles({
   },
 });
 
-export const Step = () => {
+export const Bounds = () => {
   const layoutStyles = useLayoutStyles();
   const id = useId();
 
   return (
     <div className={layoutStyles.base}>
-      <Label htmlFor={id}>Step Size</Label>
-      <SpinButton defaultValue={10} step={2} stepPage={20} id={id} />
+      <Label htmlFor={id}>Bounded SpinButton</Label>
+      <SpinButton defaultValue={10} min={0} max={20} id={id} />
+      <p>min: 0, max: 20</p>
     </div>
   );
 };
 
-Step.parameters = {
+Bounds.parameters = {
   docs: {
     description: {
-      story: `SpinButton step size can be set. Additionally \`stepPage\` can be
-      set to a large value to allow bulk steps via the \`Page Up\` and \`Page Down\` keys.`,
+      story: `SpinButton can be bounded with the \`min\` and \`max\` props.
+      Using the spin buttons or hotkeys will clamp values in the range of [min, max].
+      Users may type a value outside the range into the text input and it will not be clamped
+      by the control.`,
     },
   },
 };
