+++ packages/react-components/react-spinbutton/src/stories/SpinButton/SpinButtonDisabled.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { SpinButton } from '../index';
+import { SpinButton } from '@fluentui/react-spinbutton';
 import { Label } from '@fluentui/react-label';
 import { useId } from '@fluentui/react-utilities';
 import { tokens } from '@fluentui/react-theme';
@@ -17,26 +17,22 @@ const useLayoutStyles = makeStyles({
   },
 });
 
-export const Bounds = () => {
+export const Disabled = () => {
   const layoutStyles = useLayoutStyles();
   const id = useId();
 
   return (
     <div className={layoutStyles.base}>
-      <Label htmlFor={id}>Bounded SpinButton</Label>
-      <SpinButton defaultValue={10} min={0} max={20} id={id} />
-      <p>min: 0, max: 20</p>
+      <Label htmlFor={id}>Disabled</Label>
+      <SpinButton disabled id={id} />
     </div>
   );
 };
 
-Bounds.parameters = {
+Disabled.parameters = {
   docs: {
     description: {
-      story: `SpinButton can be bounded with the \`min\` and \`max\` props.
-      Using the spin buttons or hotkeys will clamp values in the range of [min, max].
-      Users may type a value outside the range into the text input and it will not be clamped
-      by the control.`,
+      story: `SpinButton can be disabled.`,
     },
   },
 };
