+++ packages/react-components/react-spinbutton/src/stories/SpinButton/SpinButtonUncontrolled.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { SpinButton } from '../index';
+import { SpinButton } from '@fluentui/react-spinbutton';
 import { Label } from '@fluentui/react-label';
 import { useId } from '@fluentui/react-utilities';
 import { tokens } from '@fluentui/react-theme';
@@ -17,14 +17,22 @@ const useLayoutStyles = makeStyles({
   },
 });
 
-export const Default = () => {
+export const Uncontrolled = () => {
   const layoutStyles = useLayoutStyles();
   const id = useId();
 
   return (
     <div className={layoutStyles.base}>
-      <Label htmlFor={id}>Default SpinButton</Label>
+      <Label htmlFor={id}>Uncontrolled SpinButton</Label>
       <SpinButton defaultValue={10} id={id} />
     </div>
   );
+};
+
+Uncontrolled.parameters = {
+  docs: {
+    description: {
+      story: `An uncontrolled SpinButton`,
+    },
+  },
 };
